Fix test method withBook and refactor library
