+++ src/test/java/seedu/address/logic/commands/BorrowCommandTest.java
@@ -14,7 +14,6 @@ import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.BookList;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.PersonBuilder;
