+++ src/test/java/seedu/address/storage/JsonAdaptedPersonTest.java
@@ -14,7 +14,6 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.book.Book;
 import seedu.address.model.person.Address;
-import seedu.address.model.person.BookList;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.MeritScore;
 import seedu.address.model.person.Name;
